+++ src/app/pipes/parent-child-page-path.pipe.ts
@@ -1,7 +1,7 @@
 import { Pipe, PipeTransform } from '@angular/core';
 
 import { config } from '@config';
-import { enableFrontMatterPage } from '@utility-functions';
+import { enableFrontMatterPageOrTextViewType } from '@utility-functions';
 
 
 /**
@@ -15,13 +15,13 @@ import { enableFrontMatterPage } from '@utility-functions';
 export class ParentChildPagePathPipe implements PipeTransform {
     transform(parentPath: string, childId: string): string {
         if (parentPath === '/collection') {
-            if (enableFrontMatterPage('cover', childId, config)) {
+            if (enableFrontMatterPageOrTextViewType('cover', childId, config)) {
                 return `${parentPath}/${childId}/cover`;
-            } else if (enableFrontMatterPage('title', childId, config)) {
+            } else if (enableFrontMatterPageOrTextViewType('title', childId, config)) {
                 return `${parentPath}/${childId}/title`;
-            } else if (enableFrontMatterPage('foreword', childId, config)) {
+            } else if (enableFrontMatterPageOrTextViewType('foreword', childId, config)) {
                 return `${parentPath}/${childId}/foreword`;
-            } else if (enableFrontMatterPage('introduction', childId, config)) {
+            } else if (enableFrontMatterPageOrTextViewType('introduction', childId, config)) {
                 return `${parentPath}/${childId}/introduction`;
             } else if (config.collections?.firstTextItem) {
                 const idPath = config.collections.firstTextItem[childId]?.split('_') || [];